Mar 2 dic 2025

+++ database/migrations/tenant/2025_08_13_215210_create_risk_probabilities_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('risk_probabilities', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('weight');
+            $table->decimal('weight', 5, 2);
             $table->timestamps();
         });
     }
